split test_onnxruntime.py
